Reduce number of max workers in watch mode
After running `yarn setup` or `yarn setup:expo`, running `yarn
watch:clean`, and loading the app for the first time, bundling the app
ends up hogging CPU and memory, rendering the computer useless for
anything else. (Context: The computer in question is a Mac with 16 GB of
RAM and and an M1 chip, and an average number of applications and
browser tabs are open.) This happens because Expo will start as many
workers as the number of available CPU cores.

To fix this, this PR reduces the number of workers to 3, which still
works well.